Rename sh to default, cleanup code, create executor default implementation
